net/i40e/base: improve typecasting
Adding typecast to variables to avoid compiler warnings generated on
certain platforms and compilers if variables of a particular data type
are assigned to ones of a smaller data type.

Acked-by: Bruce Richardson <bruce.richardson@intel.com>

+++ drivers/net/i40e/base/i40e_dcb.c
@@ -906,7 +906,7 @@ enum i40e_status_code i40e_init_dcb(struct i40e_hw *hw, bool enable_mib_change)
 		return I40E_ERR_NOT_READY;
 
 	/* Get the LLDP AdminStatus for the current port */
-	adminstatus = lldp_cfg.adminstatus >> (hw->port * 4);
+	adminstatus = (u8)(lldp_cfg.adminstatus >> (hw->port * 4));
 	adminstatus &= 0xF;
 
 	/* LLDP agent disabled */
